requirements.txt
pdf2image
pytesseract
tesseract
tesseract-ocr
Pillow
Flask